libgit2: update GitReference to also hold reflogs
We are using reflogs to determine which branch is created first so fetching reflogs
for the references is an everyday task to us so it makes things easier for us if we
pack reflogs together with GitReference.

We can change it to be lazy loaded afterwards if we see significant performance drop.