sparse: fix feature jumping when there is no ref
When user try to create a new feature on an empty repository using
`sparse feature` command, it was failing because there is no commit on
the repo yet. Because we were were using `start_point` (which is main by
default) with `git switch -c <new_feature_branch_name> <start_point>`
now we fixed the issue by first checking the existence of the
`start_point`. If it doesn't exists, then we are not sending a
start_point to git switch command. Then by default git uses current
branch as starting point.

Closes #32

Signed-off-by: Talha Can Havadar <havadartalha@gmail.com>