Sparse slices config (#30)
* libgit2: update GitReference to also hold reflogs

We are using reflogs to determine which branch is created first so fetching reflogs
for the references is an everyday task to us so it makes things easier for us if we
pack reflogs together with GitReference.

We can change it to be lazy loaded afterwards if we see significant performance drop.

* sparse: implement slices and use slices as backend

Created `slices.zig` which implements a function `getAllSlicesWith`
which is being used in Feature. We also changed the way we detect the
first slice. Now we use the first reflog entry of the slice to determine
the order of creation for slices.

* libgit2: add GitConfig to read config entries

* sparse: add `SparseConfig` implementation in sparse module

* sparse: do not use hardcoded email for branches

A new utility function that generates a sparse branch prefix
`sparseBranchRefPrefix` is created in utils.zig and we are using that
function to create branches.

Signed-off-by: Talha Can Havadar <havadartalha@gmail.com>

* sparse: add 'all' option for sparseBranchPrefix which ignores user

Signed-off-by: Talha Can Havadar <havadartalha@gmail.com>

* sparse: tidy up feature.zig

* sparse: slice: use const u8 for glob in getAllSlicesWith

---------

Signed-off-by: Talha Can Havadar <havadartalha@gmail.com>