libgit2: update GitReference to fetch parent branch
`createdFrom` function is implemented to fetch parent branch of the
given reference. It gets the information from reflog of the reference by
parsing the first created entry in the reflog and then tries a branch
lookup for the found reference value if it is a valid branch then
returns the reference for that branch otherwise returns false.