libgit2: remove _repo from GitReference
We dont want GitReference to hold _repo because it is not its concernt
to manage memory for repo and it adds more complexity on top of
GitReference to manage repo when we need references to be shipped
outside the libgit2 context.

If we really want to have GitReference to hold repo information we could
enforce GitReference to open the GitRepository every time it is created
but this is not what we want for now.