sparse: detect leaf slices properly
fixed memory management issues regarding GitRepository instances and
refactored the leaf slice detection logic so now we are ableto detect
which branch is created from which one. That way we have more robust way
of detecting the leaf and this also gives as flexibility when we want to
create slices in between slices.