sparse: generate slice links/graphs (#35)
* fix incorrect merge

* libgit2: update GitReference to fetch parent branch

`createdFrom` function is implemented to fetch parent branch of the
given reference. It gets the information from reflog of the reference by
parsing the first created entry in the reflog and then tries a branch
lookup for the found reference value if it is a valid branch then
returns the reference for that branch otherwise returns false.

* sparse: contruct graph for the array of slices

new `constructLinks` function in slice.zig helps creation of a slice
graph which will be useful later to determine which slice is the last
slice for the given feature.

* libgit: remove sorting helper function from GitReference

* sparse: create leafNodes function

`leafNodes` function takes slice_pool and returns the leaf slices in the
pool which can be utilized in printSliceGraph function as well.

* libgit2: free GitBranch anyways before returning from 'createdFrom'

* libgit2: remove _repo from GitReference

We dont want GitReference to hold _repo because it is not its concernt
to manage memory for repo and it adds more complexity on top of
GitReference to manage repo when we need references to be shipped
outside the libgit2 context.

If we really want to have GitReference to hold repo information we could
enforce GitReference to open the GitRepository every time it is created
but this is not what we want for now.

* sparse: detect leaf slices properly

fixed memory management issues regarding GitRepository instances and
refactored the leaf slice detection logic so now we are ableto detect
which branch is created from which one. That way we have more robust way
of detecting the leaf and this also gives as flexibility when we want to
create slices in between slices.